move shortcode function into class
